Added a slice-generator that utilizes multiprocessing.
Multiprocessing is used to parallelize the work of post-processing
individual slices. The generator is supposed to work effortlessly
with the output data-files from generate_data.py of the MLGWSC-1.

To implement post-processing, sub-class the Slicer class and
overwrite the `process_slice` method. Lighter formatting of the
data from all detectors can be performed by overwriting the
`format_returns` method.